Made changes to optimizer

+++ Optimizer.cc
@@ -461,8 +461,8 @@ QueryPlan * Optimizer::OptimizedQueryPlan() {
 
 		Attribute *attr = new Attribute[1];
 		attr[0].name = (char *)"sum";
-		attr[0].myType = Double;
-
+	//	attr[0].myType = Double;
+		attr[0].myType = Int;
 		sum->outputSchema = new Schema ((char *)"dummy", 1, attr);
 	}
 
@@ -494,7 +494,8 @@ QueryPlan * Optimizer::OptimizedQueryPlan() {
 		project->keepMe[0] = sum->outputSchema->Find((char *) "sum");
 		outputAtts = new Attribute[outputNum];
 		outputAtts[0].name = (char*)"sum";
-		outputAtts[0].myType = Double;
+		// outputAtts[0].myType = Double;
+		outputAtts[0].myType = Int;
 		ithAttr = 1;
 	}
 	else if(join) {
